server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
openai==1.3.0

